Added files via upload
created a setup file. Need to test within a diff env. 
Added missing species_request for ProjectFlask as well as flickr. 
Created a dbHandler in Pycharm to work on local db methods as well as adding data to display within the webpage. 
dbHandler needs to be incorporated into the Projectflask yet. Currently it has methods to insert entries, and then query for the description and the sciname. Currently i am using the information from GBIF, it is possible to use other descriptions provided through wikipedia api, just needs to be reconfigured to do so.